Tools_Automation/bookeep/requirements.txt
flask
line-bot-sdk
gspread
google-auth
python-dotenv
google-generativeai
gunicorn
requests
beautifulsoup4
flask-apscheduler
